+++ src/cli/addSeries.py
