+++ src/lib/serimantic/__init__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+
+"""
+serimantic
+~~~~~~~~~~
+
+*serimantic* is a NLP package working with tmdbsimple and nltk to process series
+data from TMDB.
+https://github.com/HanBnrd/Serimantic
+
+:copyright: (c) 2017-2018 by Justinien Ghorra and Johann Benerradi.
+:license: GNU GPL 3.0, see LICENSE for more details
+"""
+
+__title__ = 'serimantic'
+__version__ = '0.1'
+__author__ = 'Justinien Ghorra and Johann Benerradi'
+__copyright__ = 'Copyright (c) 2017-2018 Justinien Ghorra and Johann Benerradi'
+__license__ = 'GNU GPL 3.0'
