--- src/serimantic/__init__.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-serimantic
-~~~~~~~~~~
-
-*serimantic* is a NLP package working with tmdbsimple and nltk to process series
-data from TMDB.
-https://github.com/HanBnrd/Serimantic
-
-:copyright: (c) 2017-2018 by Serimantic.
-:license: GPLv3, see LICENSE for more details
-"""
-
-__title__ = 'serimantic'
-__version__ = '0.1'
-__author__ = 'Serimantic'
-__copyright__ = 'Copyright (c) 2017-2018 Serimantic'
-__license__ = 'GPLv3'
